Update README.md
The Biochemical Innovation Team (BIT) is a subdivision within the UBC Biomedical Engineering
Student Team and is composed of undergraduate students from the faculty of Engineering. We
are dedicated to finding novel solutions to medical challenges worldwide, with a special focus on
chemical and biological approaches.
Heart disease and similar cardiovascular diseases are the leading cause of death for adults and
those of minority groups in the entire world. Of these diseases, ischemic heart disease, or
coronary heart disease (CHD) is the most common and is often preceded by risk factors such as
high blood pressure, smoking, diabetes, lack of exercise, obesity, high blood cholesterol, poor
diet, depression, and excessive alcohol intake. Currently, there is a large gap in racial and
ethnic diversity in the field of genetic factors and CHD research. With data largely being
representative of white people, major diagnostic and risk assessment problems arise with
minority groups.
Our project aims to look into the epigenetic risk factors for various minority groups in terms of
likelihood of developing CHD in later points in life. BIT will be compiling epigenetic data from
various genetic databases and using machine learning to develop a model to predict if a
symptomatic individual has CHD. We aim to ensure this model is comprehensive, firstly in
diversity and inclusion and secondly for asymptomatic individuals that have epigenetic
precursors to the development of CHD.